[PathBundle] enhances editor
