+++ main/core/Resources/less/layout/components/summaries.less
@@ -172,7 +172,6 @@
 
     display: flex;
     align-items: center;
-    //justify-content: space-between;
 
     &:hover {
         background-color: @path-summary-link-hover-bg;
@@ -181,9 +180,6 @@
 
 .step-actions {
     white-space: nowrap; // avoid buttons to be displayed in column
-    line-height: 2em;
-    vertical-align: middle;
-    margin-left: auto; // move to the right of the flex container
 
     .btn-summary {
         padding: 2px;
@@ -238,7 +234,7 @@
         // Take full height of the viewport / content (if taller than than viewport)
         left: 0;
         top: @component-margin-bottom;
-        bottom: 0;
+        bottom: @component-margin-bottom;
 
         // hide title when closed
         .summary-title {
@@ -269,7 +265,7 @@
         // Take full height of the viewport / content (if taller than than viewport)
         left: 0;
         top: @component-margin-bottom;
-        bottom: 0;
+        bottom: @component-margin-bottom;
 
         // hide title when closed
         .summary-title {
