+++ main/core/Resources/less/layout/resource/explorer.less
@@ -25,6 +25,7 @@
 
     .summary-container {
         top: 0;
+        bottom: 0;
 
         + .content-container {
             padding-left: 0; // removes spacing between summary and list
