+++ main/core/Resources/less/themes/default/variables.less
@@ -121,7 +121,6 @@
 @breadcrumb-inner-border-size:  1;
 @breadcrumb-inner-border-color: #bbb;
 
-//@breadcrumb-link-bg:            #EAEAEA;
 @breadcrumb-link-color:         #777;
 @breadcrumb-link-hover-color:   #333;
 
@@ -200,7 +199,7 @@
 // -
 // Summaries
 //
-// As summaries can have very different forms and usage (eg. path, quiz)
+// As summaries can have very different forms and usages (eg. path, quiz)
 // there is no default implementation in the base layout
 // but each UI component intended to be used as a summary MUST use this vars
 // and adapt to the use case in order to keep common look & feel between all summary implementations
