+++ main/core/Resources/translations/platform.en.json
@@ -735,7 +735,6 @@
     "month.September": "September",
     "months": "months",
     "more_actions": "Available actions",
-    "move": "Move",
     "my_account": "My account",
     "my_agenda": "My agenda",
     "my_applications": "My applications",
@@ -1695,5 +1694,9 @@
     "colors-selector": "Selector",
     "publication_by_date": "Access period",
     "confirmation_checkbox": "By checking this box, I confirm that I have read and understood the warning.",
-    "publication_by_date_help": "Resource must be published beforehand, then define dates in format dd/mm/yyyy HH:mm. <br/>Ex: 25/12/2018 14:00"
+    "publication_by_date_help": "Resource must be published beforehand, then define dates in format dd/mm/yyyy HH:mm. <br/>Ex: 25/12/2018 14:00",
+    "movement": "Move",
+    "before": "Before",
+    "after": "After",
+    "deletion": "Deletion"
 }
