+++ plugin/path/Resources/modules/resources/path/editor/modals/position/store/index.js
@@ -1,0 +1,8 @@
+
+import {reducer} from '#/plugin/path/resources/path/editor/modals/position/store/reducer'
+import {selectors} from '#/plugin/path/resources/path/editor/modals/position/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
