+++ plugin/path/Resources/modules/resources/path/editor/modals/position/store/reducer.js
@@ -1,0 +1,9 @@
+import {makeFormReducer} from '#/main/app/content/form/store/reducer'
+
+import {selectors} from '#/plugin/path/resources/path/editor/modals/position/store/selectors'
+
+const reducer = makeFormReducer(selectors.STORE_NAME)
+
+export {
+  reducer
+}
