+++ plugin/path/Resources/modules/resources/path/editor/modals/position/store/selectors.js
@@ -1,0 +1,6 @@
+
+const STORE_NAME = 'stepPositionForm'
+
+export const selectors = {
+  STORE_NAME
+}
