+++ plugin/path/Resources/modules/resources/path/overview/components/main.jsx
@@ -1,18 +1,16 @@
 import React from 'react'
-import {connect} from 'react-redux'
 import {PropTypes as T} from 'prop-types'
 
 import {trans} from '#/main/app/intl/translation'
 import {LINK_BUTTON} from '#/main/app/buttons'
+
 import {ResourceOverview} from '#/main/core/resource/components/overview'
-import {selectors as resourceSelectors} from '#/main/core/resource/store/selectors'
 import {UserEvaluation as UserEvaluationTypes} from '#/main/core/resource/prop-types'
 
-import {Summary} from '#/plugin/path/resources/path/overview/components/summary'
-import {selectors} from '#/plugin/path/resources/path/store'
+import {OverviewSummary} from '#/plugin/path/resources/path/overview/components/summary'
 import {Path as PathTypes} from '#/plugin/path/resources/path/prop-types'
 
-const OverviewComponent = props =>
+const OverviewMain = props =>
   <ResourceOverview
     contentText={props.path.display.description}
     progression={{
@@ -38,13 +36,13 @@ const OverviewComponent = props =>
     <section className="resource-parameters">
       <h3 className="h2">{trans('summary')}</h3>
 
-      <Summary
+      <OverviewSummary
         steps={props.path.steps}
       />
     </section>
   </ResourceOverview>
 
-OverviewComponent.propTypes = {
+OverviewMain.propTypes = {
   path: T.shape(
     PathTypes.propTypes
   ).isRequired,
@@ -54,19 +52,11 @@ OverviewComponent.propTypes = {
   )
 }
 
-OverviewComponent.defaultProps = {
+OverviewMain.defaultProps = {
   empty: true,
   evaluation: {}
 }
 
-const Overview = connect(
-  (state) => ({
-    path: selectors.path(state),
-    empty: selectors.empty(state),
-    evaluation: resourceSelectors.resourceEvaluation(state)
-  })
-)(OverviewComponent)
-
 export {
-  Overview
+  OverviewMain
 }
