+++ plugin/path/Resources/styles/resources/path/navigation.less
@@ -6,6 +6,7 @@
     display: flex;
     flex-direction: row;
     align-items: stretch;
+    margin-bottom: @component-margin-bottom;
 
     .btn-previous,
     .btn-next {
