+++ plugin/path/Resources/styles/resources/path/step.less
@@ -12,11 +12,17 @@
     }
 }
 
+.parameters-title,
 .step-title {
     display: flex;
     flex-direction: row;
     align-items: center;
     justify-content: left;
+}
+
+.step-toolbar {
+    margin-left: auto;
+    white-space: nowrap;
 }
 
 .step-manual-progression {
