+++ plugin/path/Resources/styles/variables.less
@@ -5,7 +5,8 @@
 // Claroline & bootstrap variables
 @import "../../../../main/core/Resources/less/themes/default/variables";
 
-@path-summary-lvl-indent:           20px; // indent children of a step in the summary
+
+@path-summary-lvl-indent: 20px; // indent children of a step in the summary
 
 
 // -
