+++ plugin/path/Resources/translations/path.en.json
@@ -9,8 +9,6 @@
     "open_summary": "Open summary",
     "path_manual_progression_allowed": "Allow user to mark its progression",
     "primary_resource": "Primary resource",
-    "resource_delete_confirm": "Are you sure you want to delete this resource ?",
-    "resource_delete_title": "Delete resource",
     "resource_hide_parent": "Hide in this step",
     "resource_no_propagate": "Make unavailable in children",
     "resource_propagate": "Make available in children",
@@ -26,7 +24,6 @@
     "step_content_height": "Height for step content",
     "step_content_height_info": "To use only if the content of the step doesn't display correctly. (0 = disabled)",
     "step_delete_confirm": "Are you sure you want to delete this step ?",
-    "step_delete_title": "Delete step",
     "user_progression": "status: ",
     "user_progression_step_done": "You have marked this step as done.",
     "user_progression_step_done_label": "Done",
