+++ plugin/path/Resources/translations/path.fr.json
@@ -9,8 +9,6 @@
     "open_summary": "Ouvrir le sommaire",
     "path_manual_progression_allowed": "Permettre à l'utilisateur d'indiquer manuellement sa progression ?",
     "primary_resource": "Ressource principale",
-    "resource_delete_confirm": "Êtes-vous sur de vouloir supprimer cette ressource ?",
-    "resource_delete_title": "Supprimer la ressource",
     "resource_hide_parent": "Masquer dans cette étape",
     "resource_no_propagate": "Rendre indisponible dans les étapes enfants",
     "resource_propagate": "Rendre disponible dans les étapes enfants",
@@ -26,7 +24,6 @@
     "step_content_height": "Taille pour le contenu de l'étape",
     "step_content_height_info": "À utiliser que si le contenu de l'étape ne s'affiche pas correctement. (0 = désactivé)",
     "step_delete_confirm": "Êtes-vous sur de vouloir supprimer cette étape (ainsi que toutes les étapes enfants associées) ?",
-    "step_delete_title": "Supprimer l'étape",
     "user_progression": "statut : ",
     "user_progression_step_done": "Vous avez marqué cette étape comme étant terminée.",
     "user_progression_step_done_label": "Fait",
